% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperMPLE.R
\name{compute_graphStructure}
\alias{compute_graphStructure}
\title{Pre-computes graph structures of the point pattern and the integration grid to speed up the PL optimisation.}
\usage{
compute_graphStructure(
  pointPattern,
  interactionRadius,
  hardCoreRadius,
  trend,
  gridSizeX,
  gridSizeY,
  isHomogeneous
)
}
\arguments{
\item{pointPattern}{the considered point pattern.}

\item{interactionRadius}{the interaction radius $R > r$ of the Strauss process.}

\item{hardCoreRadius}{the hard core distance $r>0$ of the Strauss process.}

\item{trend}{the trend/ activation function of the Strauss process. Might be a real number or a pixel image.}

\item{gridSizeX}{the number of grid points to compute the normalisation constant of the pseudo-likelihood in x-direction.}

\item{gridSizeY}{the number of grid points to compute the normalisation constant of the pseudo-likelihood in y-direction.}
}
\value{
A list of containing the pre-computed graph structures of the point pattern and the integration grid.
}
\description{
Pre-computes graph structures of the point pattern and the integration grid to speed up the PL optimisation.
}
\author{
Johannes Wieditz
}
