% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperMPLE.R
\name{conditionalIntensity}
\alias{conditionalIntensity}
\title{Computes the conditional intensity of a given point pattern stemming from a PoSt-hard core process.}
\usage{
conditionalIntensity(theta, graphStructure, i, trend, isHomogeneous)
}
\arguments{
\item{theta}{the model parameters.}

\item{graphStructure}{the interaction graph induced by the point pattern as computed by the function compute_graphStructure (1st component).}

\item{i}{the index of the point of the point pattern considered.}

\item{trend}{the trend/ activation function of the Strauss process. Might be a real number or a pixel image.}

\item{isHomogeneous}{a flag indicating whether the considered process is homogeneous (TRUE) or not (FALSE)}
}
\value{
The conditional intensity of the i-th point in the point pattern listed in the graph structure passed.
}
\description{
Computes the conditional intensity of a given point pattern stemming from a PoSt-hard core process.
}
\author{
Johannes Wieditz
}
