% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperMPLE.R
\name{logPseudoLikelihood}
\alias{logPseudoLikelihood}
\title{Computes the log pseudo-likelihood of a given point pattern stemming from a PoSt-hard core process.}
\usage{
logPseudoLikelihood(
  theta,
  graphStructure,
  integralGridStructure,
  trend,
  isHomogeneous,
  windowVolume
)
}
\arguments{
\item{theta}{the model parameters.}

\item{graphStructure}{the interaction graph induced by the point pattern as computed by the function compute_graphStructure (1st component).}

\item{integralGridStructure}{the interaction graph induced by the point pattern and the integration grid as computed by the function compute_graphStructure (2nd component).}

\item{trend}{the trend/ activation function of the Strauss process. Might be a real number or a pixel image.}

\item{isHomogeneous}{a flag indicating whether the considered process is homogeneous (TRUE) or not (FALSE)}
}
\value{
The conditional intensity of the i-th point in the point pattern listed in the graph structure passed.
}
\description{
Computes the log pseudo-likelihood of a given point pattern stemming from a PoSt-hard core process.
}
\author{
Johannes Wieditz
}
