% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperMPLE.R
\name{SuperMPLE}
\alias{SuperMPLE}
\title{Computes a maximum pseudo-likelihood estimator for the passed point pattern in the PoSt-hard core model.}
\usage{
SuperMPLE(
  pointPattern,
  interactionRadius,
  hardCoreRadius,
  trend,
  gridSizeX,
  gridSizeY,
  theta_init,
  isHomogeneous
)
}
\arguments{
\item{pointPattern}{the considered point pattern.}

\item{interactionRadius}{the interaction radius $R > r$ of the Strauss process.}

\item{hardCoreRadius}{the hard core distance $r>0$ of the Strauss process.}

\item{trend}{the trend/ activation function of the Strauss process. Might be a real number or a pixel image.}

\item{gridSizeX}{the number of grid points to compute the normalisation constant of the pseudo-likelihood in x-direction.}

\item{gridSizeY}{the number of grid points to compute the normalisation constant of the pseudo-likelihood in y-direction.}

\item{theta_init}{intial value where the optimisation procedure starts from.}
}
\value{
A list of containing the pre-computed graph structures of the point pattern and the integration grid. Note, that the parameters are stored in the natural parameter space, i.e. on log-scale.
}
\description{
Computes a maximum pseudo-likelihood estimator for the passed point pattern in the PoSt-hard core model.
}
\examples{
\donttest{
interactionRadius <- 0.1
hardCoreRadius    <- 0.03
trend <- 42
set.seed(42)
straussPoints <- rStraussHard(beta = 1 * trend, gamma = 0.4, R = interactionRadius, H = hardCoreRadius)
poissonPoints <- rpoispp(lambda = 12)
pointPattern  <- superimpose(straussPoints, poissonPoints)
gridSizeX     <- gridSizeY <- 10
theta_init    <- c( beta = log(1), gamma = log(.4), lambda = log(12) )
MPLE          <- SuperMPLE(pointPattern, interactionRadius, hardCoreRadius, trend, gridSizeX, gridSizeY, theta_init, TRUE)
MPLE
}


}
\author{
Johannes Wieditz
}
